+++ .vscode/settings.json
@@ -16,6 +16,7 @@
     "Arbitrum",
     "basegoerli",
     "bignumber",
+    "cxtoken",
     "defi",
     "divu",
     "Gargantuworm",
