+++ src/nft-minter/merkle/interfaces/IMerkleProofMinter.sol
@@ -16,6 +16,7 @@ interface IMerkleProofMinter {
   event MintedWithProof(address indexed account, bytes32[] proof, uint256 level, uint256 tokenId);
   event PersonaSet(address indexed account, uint8 level, uint8 persona);
 
+  error InsufficientNpmBalanceError(uint256 minRequired);
   error InvalidBindingError(address account, uint256 boundTokenId);
   error InvalidPersonaError();
   error PersonaAlreadySetError();
